exalt: Add `Common.hpp` to hold shared type aliases.
